+++ spec/isolation/components/mailer_configuration/with_hanami_mailer_spec.rb
@@ -8,5 +8,22 @@ RSpec.describe "Components: mailer.configuration", type: :cli do
         expect(Hanami::Components['mailer.configuration']).to be_kind_of(Hanami::Mailer::Configuration)
       end
     end
+
+    it "resolves mailer configuration for current environment" do
+      with_project do
+        unshift "config/environment.rb", "ENV['HANAMI_ENV'] = 'production'"
+        write "script/components", <<-EOF
+require "\#{__dir__}/../config/environment"
+Hanami::Components.resolve('mailer.configuration')
+
+configuration = Hanami::Components['mailer.configuration']
+puts "mailer.configuration.delivery_method: \#{configuration.delivery_method.first.inspect}"
+EOF
+
+        bundle_exec "ruby script/components"
+
+        expect(out).to include("mailer.configuration.delivery_method: :smtp")
+      end
+    end
   end
 end